fix(audio): widen the class-weight denominator in AudioClassifierBase
CodeQL cs/loss-of-precision (alerts 10631, 10632), AudioClassifierBase.cs:232-238:
numClasses * count was an int product; 1000 classes x 2.2M samples wraps and yields a negative
class weight.

+++ src/Audio/Classification/AudioClassifierBase.cs
@@ -229,13 +229,13 @@ public abstract class AudioClassifierBase<T> : AudioNeuralNetworkBase<T>,
             if (count > 0)
             {
                 // Inverse frequency weighting
-                double weight = (double)totalSamples / (numClasses * count);
+                double weight = (double)totalSamples / ((double)numClasses * count);
                 weights[label] = NumOps.FromDouble(weight);
             }
             else
             {
                 // Assign maximum weight to classes with zero samples (they need the most attention)
-                double maxWeight = nonZeroCounts.Max(kvp => (double)totalSamples / (numClasses * kvp.Value));
+                double maxWeight = nonZeroCounts.Max(kvp => (double)totalSamples / ((double)numClasses * kvp.Value));
                 weights[label] = NumOps.FromDouble(maxWeight);
             }
         }
